* newDesign: (498 commits)
  Info.plist update
  Removed unused classes
  Disabled iPad native screen resolution
  Quickblox iOS SDK upd
  Added necessary internet connection checks
  Quickblox iOS SDK 2.7.3
  Password hint fix
  Removed weakify from animatewithduration
  Removed landscape for now
  Status bar appearance for calls controller
  Showing hint if connection lost
  Caller name in call push notification
  Updated Digits api keys
  Auto resizing mask for message notification
  Quickblox iOS SDK 2.7.2.3
  Fixed push notifications not being subscribed after QMProfile was cleared incorrectly
  Updating global search table after contact list did change to display correct value for add button
  Updated README.md (was removed by wrong merge)
  removed reload row that caused crash cause table view based on reload data
  removed reloadrow caused crash after another delegate called reloaddata
  ...

# Conflicts:
#	Q-municate/Classes/ViewControllers/TabBarController/CallControllers/QMVideoCallController/QMVideoCallController.m
#	Q-municate/Classes/ViewControllers/TabBarController/CallControllers/QMVideoP2PController/QMVideoP2PController.m
#	Q-municate/Main.storyboard

+++ Q-municate/Categories/UIImage+fixOrientation.m
@@ -23,19 +23,19 @@
         case UIImageOrientationDown:
         case UIImageOrientationDownMirrored:
             transform = CGAffineTransformTranslate(transform, self.size.width, self.size.height);
-            transform = CGAffineTransformRotate(transform, M_PI);
+            transform = CGAffineTransformRotate(transform, (CGFloat)M_PI);
             break;
             
         case UIImageOrientationLeft:
         case UIImageOrientationLeftMirrored:
             transform = CGAffineTransformTranslate(transform, self.size.width, 0);
-            transform = CGAffineTransformRotate(transform, M_PI_2);
+            transform = CGAffineTransformRotate(transform, (CGFloat)M_PI_2);
             break;
             
         case UIImageOrientationRight:
         case UIImageOrientationRightMirrored:
             transform = CGAffineTransformTranslate(transform, 0, self.size.height);
-            transform = CGAffineTransformRotate(transform, -M_PI_2);
+            transform = CGAffineTransformRotate(transform, (CGFloat)-M_PI_2);
             break;
         case UIImageOrientationUp:
         case UIImageOrientationUpMirrored:
@@ -63,7 +63,7 @@
     
     // Now we draw the underlying CGImage into a new context, applying the transform
     // calculated above.
-    CGContextRef ctx = CGBitmapContextCreate(NULL, self.size.width, self.size.height,
+    CGContextRef ctx = CGBitmapContextCreate(NULL, (size_t)self.size.width, (size_t)self.size.height,
                                              CGImageGetBitsPerComponent(self.CGImage), 0,
                                              CGImageGetColorSpace(self.CGImage),
                                              CGImageGetBitmapInfo(self.CGImage));
